Create README.md
we have data from people with heart disease and people without. we need to clean them up and prepare them. 
we have to see what characteristics are more related to the likelihood of someone having it.
finally we must build a model that predicts whether a person has a disease or not, but mainly we are interested in the model to identify those who have it because it is the most important in medical data.